Feature/tela inicial

+++ app/src/main/java/com/example/cotacaomobile/ConvMoeda.kt
@@ -1,7 +1,7 @@
 package com.example.cotacaomobile
 
 data class ConvMoeda(
-    val real: Float,
-    val dolar: Float,
-    val bitcoin: Float
+    val real: Double,
+    val dolar: Double,
+    val bitcoin: Double
 )
